banking-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.29
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
httpx==0.26.0
email-validator==2.1.0.post1